refactor book1js to use createSlice
