+++ src/redux/configureStore.js
@@ -1,11 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
 import statusReducer from './categories/categories';
-import reducer from './books/books';
+import bookReducer from './books/book1';
 
 const store = configureStore({
   reducer: {
     status: statusReducer,
-    book: reducer,
+    book: bookReducer,
   },
 });
